refactor(vue): extract drawing interaction into useDrawingManager composable
Move drawingController, selectedDrawing state, and all drawing handlers
(handleSelectTool, onUpdateDrawingStyle, onDeleteDrawing) into
packages/vue/src/composables/chart/useDrawingManager.ts.
Reduces KLineChart.vue <script setup> by ~50 lines.
